requirements.txt
cycler==0.10.0
decorator==4.3.0
kiwisolver==1.0.1
lime==0.1.1.31
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
opencv-python==3.4.0.12
pathlib==1.0.1
Pillow==5.1.0
pkg-resources==0.0.0
pybind11==2.2.3
pyparsing==2.2.0
pypng==0.0.18
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
wisardpkg==0.0.5
